Extract cursor related DSL methods to Cursors::DSL.
