+++ lib/vedeu/all.rb
@@ -27,11 +27,11 @@ require 'vedeu/null/all'
 require 'vedeu/models/views/all'
 require 'vedeu/models/all'
 require 'vedeu/application/all'
+require 'vedeu/cursors/all'
 require 'vedeu/dsl/all'
 require 'vedeu/borders/all'
 require 'vedeu/buffers/all'
 require 'vedeu/colours/all'
-require 'vedeu/cursors/all'
 require 'vedeu/distributed/all'
 require 'vedeu/editor/all'
 require 'vedeu/geometry/all'
